- Introduced `ShiftPettyCash` command for managing petty cash shifts.
- Created `ReportsController` for handling various report generation.
- Added `myBranches` method in `AdminController` to retrieve branches associated with the authenticated admin.
- Updated various controllers to enhance functionality and improve response formats.
- Added new fields in models and migrations to support additional data requirements.
- Enhanced frontend components for better user experience and interaction.

+++ database/migrations/2025_01_10_015000_create_payments_table.php
@@ -19,8 +19,12 @@ return new class extends Migration
             $table->string('WalkInName')->nullable();    // for walk-in clients
             $table->string('BookingRef')->nullable();     // if referencing a booking
             $table->string('SessionRef')->nullable();     // if referencing a session 
-            $table->unsignedBigInteger('MonthlyClientID')->nullable();
-            $table->foreign('MonthlyClientID')->references('MonthlyClientID')->on('monthly_clients')->onDelete('cascade'); // or set null
+            $table->unsignedInteger('MonthlyClientID')->nullable();
+            $table
+              ->foreign('MonthlyClientID')
+              ->references('MonthlyClientID')
+              ->on('monthly_clients')
+              ->onDelete('cascade');            
             $table->text('PaymentFor')->nullable();          
             $table->string('PaymentMethod')->nullable();    // "Cash", "GCash", "BPI"
             $table->decimal('Amount', 10, 2);
